feat: v2.619.0 智能磁盘健康预测系统/CI修复
新功能:
- 智能磁盘健康预测 (diskhealth)
  - S.M.A.R.T. 数据实时采集与分析
  - AI 驱动的磁盘故障预测
  - 健康评分系统 (0-100)
  - 多级告警机制
  - 风险因素识别
  - RESTful API

修复:
- tiering: TierTypeMemory 未定义
- tiering: WarmThreshold 缺失字段
- ups: CreateShutdownTask 返回值处理
- integration: int/int64 类型转换

+++ tests/integration/tiering_test.go
@@ -131,7 +131,7 @@ func (m *MockTieringManager) Migrate(ctx context.Context, req *tiering.MigrateRe
 		SourceTier:  req.SourceTier,
 		TargetTier:  req.TargetTier,
 		Action:      req.Action,
-		TotalFiles:  int64(len(req.Paths)),
+		TotalFiles:  len(req.Paths),
 		TotalBytes:  1000000000,
 	}
